refactor(lib): extract addon-availability host-capability probes from addons.ts
Move the host-capability detection (voiceImageRef, dockerManifestExists,
probeCuda, probeRocm, getAddonProfileAvailability, the AddonProfileAvailability
type, the process-lifetime cache, and the execFile wrapper) out of addons.ts
into a dedicated control-plane/addon-availability.ts. These probe the host
(docker runtimes, GPU device nodes, published image manifests) rather than
compose/state, so they form their own cohesive module.

Because the probes are now testable in isolation, the ugly
`__addonAvailabilityTestHooks` internal-export is REMOVED: the cache reset and
execFile wrapper are exported directly from the new module as
`resetAvailabilityCache` and `execFileNoThrow`, and addons.test.ts imports them
straight from addon-availability.js. addons.ts keeps annotateAddonProfileAvailability
(which now imports getAddonProfileAvailability from the new module). The barrel
re-exports the moved symbols so consumers (CLI, UI) are unaffected.
